pyproject.toml
[tool.poetry]
name = "telemetry-generation"
version = "0.1.0"
description = ""
authors = ["Raymond Bernardo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.2"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
